fix(build): a clean build, under both compilers
The build worked and said a great deal while doing it: 2141 warnings
under gcc, 4 under clang, and one target that did not build at all on a
machine that had run `make setup`. None of them pointed at a defect,
which is the problem - a build nobody can read is a build where the next
warning goes unread too.

The flood was one cause. inkcell's `inkcell_str_id` was an enumeration
type naming inkcell's own fifteen ids, and this client's ~870 continue
that numbering in an enum of their own, so every id this client names -
every `inkcell_str()` call, every `.label =` in a table - was a
conversion between two enumeration types that nothing says agree. gcc is
right about that and said so 2094 times; clang happens not to warn
between anonymous and named enums, which is why it was never seen. The
answer is upstream in inkcell, where the id type is now the index it
always was, and the follow-through here is the spelling: `enum
inkcell_str_id` becomes `inkcell_str_id`. No behaviour changes - an id
was already an int-width table index into the one catalog
mesh_i18n_register() hands over.

The rest, one each:

  - ble_hci.c defined _DEFAULT_SOURCE bare while inkwell -D's it for
    everything that links it, and the two definitions are not the same
    definition. Guarded.
  - Two locals left in the compose screen's keyboard renderer after the
    text field moved into inkcell. Deleted.
  - nav_dialog_answer() narrowed an unsigned ternary into a uint8_t
    without saying so. Cast.
  - scripts/linux-cli-build.sh built a static binary against whatever
    SDL2 the host had, so on a developer's own machine - where `make
    setup` installs libsdl2-dev - the link died on "attempted static
    link of dynamic object", while CI, which installs none for that job,
    stayed green. INKCELL_WITH_SDL=OFF exists for exactly this; the
    cross build says it too rather than relying on the container
    happening to carry no aarch64 SDL2.

Verified clean and green under gcc and clang, Debug and Release, with
and without SDL2 present, under ASan+UBSan, and for the static musl CLI.

+++ src/ui/nav/nav.c
@@ -2669,8 +2669,8 @@ bool mesh_ui_nav_close_verify_number(struct mesh_ui_nav *nav) {
  */
 uint8_t mesh_ui_nav_dialog_answer(const struct mesh_ui_store *store, enum inkcell_key key,
                                   uint8_t cursor) {
-    const uint8_t here = cursor == 0U ? 0U : 1U;
-    const uint8_t other = here == 0U ? 1U : 0U;
+    const uint8_t here = (uint8_t)(cursor == 0U ? 0U : 1U);
+    const uint8_t other = (uint8_t)(here == 0U ? 1U : 0U);
     const struct inkcell_focus_map *const map = store != NULL ? store->focus : NULL;
     enum inkcell_focus_dir dir;
     if (map == NULL || !inkcell_focus_dir_for_key(key, &dir)) {
